Update to two-host-model
Named values for the inits so that they are easier to interpret, and then these new names can be used when plotting for more clarity. 
Totals were added to the out data frame so that it is easier to check that the model is behaving as it should. 
Figure legends were added to the plots for more clarity on what line represents what value.